moloch v2 construction require tests
